Refactor image problems into separate files
PiperOrigin-RevId: 182851572
