+++ tensor2tensor/models/slicenet_test.py
@@ -23,7 +23,7 @@ from __future__ import print_function
 
 import numpy as np
 
-from tensor2tensor.data_generators import image  # pylint: disable=unused-import
+from tensor2tensor.data_generators import cifar  # pylint: disable=unused-import
 from tensor2tensor.layers import modalities  # pylint: disable=unused-import
 from tensor2tensor.models import slicenet
 from tensor2tensor.utils import registry
